server/requirements.txt
fastapi
uvicorn[standard]
yt-dlp
google-generativeai
supabase
python-dotenv
python-multipart
pydantic
httpx
